requirements.txt
python-dotenv
matplotlib
numpy
opencv-python
opencv-python-headless
pandas
pandocfilters
scikit-image
scikit-learn
scipy
pillow
plotly
tensorboard
torch
torchaudio
torchinfo
torchmetrics
torchsummary
torchvision
tzdata
watchdog